+++ vrchat_dart/lib/src/model/api/vrc_response.dart
@@ -1,3 +1,4 @@
+// Package imports:
 import 'package:dio/dio.dart';
 
 /// A response from the VRChat API