+++ vrchat_dart/lib/src/model/streaming/streamed_current_user.dart
@@ -1,5 +1,8 @@
+// Package imports:
 import 'package:json_annotation/json_annotation.dart';
 import 'package:vrchat_dart_generated/vrchat_dart_generated.dart';
+
+// Project imports:
 import 'package:vrchat_dart/src/convenience/serializers.dart';
 
 part 'streamed_current_user.g.dart';