+++ vrchat_dart/lib/src/model/streaming/vrc_streaming_event.dart
@@ -1,7 +1,10 @@
+// Package imports:
 import 'package:json_annotation/json_annotation.dart';
 import 'package:vrchat_dart_generated/vrchat_dart_generated.dart';
+
+// Project imports:
 import 'package:vrchat_dart/src/convenience/serializers.dart';
-import 'package:vrchat_dart/src/streaming/streamed_current_user.dart';
+import 'package:vrchat_dart/src/model/streaming/streamed_current_user.dart';
 
 part 'vrc_streaming_event.g.dart';