+++ vrchat_dart/lib/vrchat_api_generator_config.dart
@@ -1,3 +1,4 @@
+// Package imports:
 import 'package:openapi_generator_annotations/openapi_generator_annotations.dart';
 
 @Openapi(